moved duplicated styles to one file
