--- src/pages/ProfilePage/Tabs/Collectibles/Collectibles.module.scss
@@ -1,39 +1,0 @@
-@import '../../../../styles/helpers.scss';
-.page_body__top {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 32px;
-  &_sorters {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 12px;
-  }
-}
-
-.title {
-  margin-bottom: 5px;
-}
-
-.counter {
-  font-size: 16px;
-  color: $color-neutral-20;
-}
-
-.artworks {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 25px;
-  margin-bottom: 32px;
-  justify-items: center;
-}
-
-.button_more {
-  padding: 0 32px;
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.button_wrapper {
-  display: flex;
-  justify-content: center;
-}
