+++ src/pages/ProfilePage/Tabs/TabHeader/TabHeader.module.scss
@@ -1,4 +1,5 @@
 @import '../../../../styles/helpers.scss';
+
 .page_body__top {
   display: flex;
   justify-content: space-between;
@@ -17,23 +18,4 @@
 .counter {
   font-size: 16px;
   color: $color-neutral-20;
-}
-
-.artworks {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 25px;
-  margin-bottom: 32px;
-  justify-items: center;
-}
-
-.button_more {
-  padding: 0 32px;
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.button_wrapper {
-  display: flex;
-  justify-content: center;
 }
